refactor: useControlQuantity 함수 네이밍 변경
